reset maxsize to 100, remove some comments

+++ main.py
@@ -18,7 +18,7 @@ def main():
     main_board.debug()
 
     policyLearner = PolicyLearner(main_board, original_player_loc, original_walls_loc, original_boxes_loc, original_terminal_loc)
-    policyLearner.learn(int(sys.argv[2]))
+    policyLearner.learn(float(sys.argv[2]))
 
 if __name__ == "__main__":
     main()
